+++ source/extensions/filters/http/jwt_authn/matcher.cc
@@ -1,6 +1,7 @@
 #include "extensions/filters/http/jwt_authn/matcher.h"
 
 #include "common/common/logger.h"
+#include "common/common/matchers.h"
 #include "common/common/regex.h"
 #include "common/router/config_impl.h"
 
@@ -58,13 +59,12 @@ private:
 class PrefixMatcherImpl : public BaseMatcherImpl {
 public:
   PrefixMatcherImpl(const RequirementRule& rule)
-      : BaseMatcherImpl(rule), prefix_(rule.match().prefix()) {}
+      : BaseMatcherImpl(rule), prefix_(rule.match().prefix()),
+        path_matcher_(Matchers::PathMatcher::createPrefix(prefix_, !case_sensitive_)) {}
 
   bool matches(const Http::HeaderMap& headers) const override {
     if (BaseMatcherImpl::matchRoute(headers) &&
-        (case_sensitive_
-             ? absl::StartsWith(headers.Path()->value().getStringView(), prefix_)
-             : absl::StartsWithIgnoreCase(headers.Path()->value().getStringView(), prefix_))) {
+        path_matcher_->match(headers.Path()->value().getStringView())) {
       ENVOY_LOG(debug, "Prefix requirement '{}' matched.", prefix_);
       return true;
     }
@@ -74,6 +74,7 @@ public:
 private:
   // prefix string
   const std::string prefix_;
+  const Matchers::PathMatcherConstSharedPtr path_matcher_;
 };
 
 /**
@@ -82,19 +83,14 @@ private:
 class PathMatcherImpl : public BaseMatcherImpl {
 public:
   PathMatcherImpl(const RequirementRule& rule)
-      : BaseMatcherImpl(rule), path_(rule.match().path()) {}
+      : BaseMatcherImpl(rule), path_(rule.match().path()),
+        path_matcher_(Matchers::PathMatcher::createExact(path_, !case_sensitive_)) {}
 
   bool matches(const Http::HeaderMap& headers) const override {
-    if (BaseMatcherImpl::matchRoute(headers)) {
-      const Http::HeaderString& path = headers.Path()->value();
-      const size_t compare_length =
-          path.getStringView().length() - Http::Utility::findQueryStringStart(path).length();
-      auto real_path = path.getStringView().substr(0, compare_length);
-      bool match = case_sensitive_ ? real_path == path_ : StringUtil::caseCompare(real_path, path_);
-      if (match) {
-        ENVOY_LOG(debug, "Path requirement '{}' matched.", path_);
-        return true;
-      }
+    if (BaseMatcherImpl::matchRoute(headers) &&
+        path_matcher_->match(headers.Path()->value().getStringView())) {
+      ENVOY_LOG(debug, "Path requirement '{}' matched.", path_);
+      return true;
     }
     return false;
   }
@@ -102,6 +98,7 @@ public:
 private:
   // path string.
   const std::string path_;
+  const Matchers::PathMatcherConstSharedPtr path_matcher_;
 };
 
 /**